fix encoding, remove header row, importing still broken

+++ pg_importer.py
@@ -1,7 +1,7 @@
 import psycopg2
 import os
-import unicodecsv as csv
-# import csv
+# import unicodecsv as csv
+import csv
 
 
 
@@ -27,18 +27,20 @@ def main():
     for year in ['2012', '2016']:
         for filename in os.listdir(fileDir + '/' + year):
 
-            # first import file into temp table
-            # copyIntoPG(filename)
-            cursor = pgConn.cursor()
-            cursor.execute("delete from temp;")
-            importFile = open(fileDir + '/' + year + '/' + filename, "r+", encoding='utf-8')
-            # importFile = open(fileDir + '/' + year + '/' + filename, "r+")
-            cursor.copy_from(importFile, 'temp', sep = ',')
-            cursor.commit()
+            if filename.endswith('.csv'):
+                # first import file into temp table
+                # copyIntoPG(filename)
+                print("filename: " + fileDir + '/' + year + '/' + filename)
+                cursor = pgConn.cursor()
+                cursor.execute("delete from temp;")
+                # importFile = open(fileDir + '/' + year + '/' + filename, "r+", encoding='utf-8')
+                importFile = open(fileDir + '/' + year + '/' + filename, "r+")
+                cursor.copy_from(importFile, 'temp', sep = ',')
+                cursor.commit()
 
-            # merge temp table into primary table
-            # cursor.execute("delete from exit_polls where year = " + year + " and state = " + state + ";")
-            # cursor.execute("insert into exit_polls select * from temp;")
+                # merge temp table into primary table
+                # cursor.execute("delete from exit_polls where year = " + year + " and state = " + state + ";")
+                # cursor.execute("insert into exit_polls select * from temp;")
